src/util/ConfigManager/ValidatorConcept: Moved the concept to ConfigManager.h.
